Change "font" and "position" managment
Issue : none

Before, to retrieve "font" and "position" infos, data was computed from
the text edit zone coordonate.
Now "font" and "position" infos are retrieved directly from the subtitle
database if a subtitle is indexed for the current time.
If not, the default infos are retrieve.

When a "font" or "position" parameter are changed from tool boxes:
if there are no subtitle indexed for the current time, the parameter become
the default parameter.

Signed-off-by: juiice <decaen.julien@gmail.com>